refactor(access): 重构反潜回模块依赖关系，消除代码重复
- 修改AccessVerificationManager构造函数，注入AntiPassbackManager替代AntiPassbackRecordDao
- 删除AccessVerificationManager中约200行重复代码（verifyAntiPassback、recordAntiPassback、getAntiPassbackConfig、getLastRecordWithCache、AntiPassbackConfig内部类）
- 修改AccessVerificationManager方法实现，委托给AntiPassbackManager统一管理
- 更新AccessManagerConfiguration，使用@DependsOn确保Bean注册顺序正确
- 在EdgeVerificationStrategy中添加反潜回记录调用（边缘验证模式：只记录，不验证）
- 在AccessRecordBatchService中添加批量反潜回记录调用（边缘验证模式：只记录，不验证）

业务逻辑说明：
- 后台验证模式：软件端验证反潜回（可跨设备）+ 记录
- 边缘验证模式：软件端只记录反潜回信息（不验证，设备端已验证），用于统计和分析

架构改进：
- 消除代码重复，符合DRY原则
- 统一依赖关系，所有反潜回相关调用都通过AntiPassbackManager
- 符合单一职责、依赖倒置原则
- 提高可维护性和可扩展性

+++ microservices/microservices-common-permission/src/main/java/net/lab1024/sa/common/permission/monitor/PermissionPerformanceMonitor.java
@@ -536,8 +536,8 @@ public class PermissionPerformanceMonitor {
         private Long minValidationTime = 0L;
         private Double cacheHitRate = 0.0;
         private Double averageCacheTime = 0.0;
-        private Integer localCacheSize = 0;
-        private Integer redisCacheSize = 0;
+        private Long localCacheSize = 0L;
+        private Long redisCacheSize = 0L;
         private Double localHitRate = 0.0;
         private Double redisHitRate = 0.0;
         private Double overallHitRate = 0.0;
